backend/pyproject.toml
[tool.poetry]
name = "hr-backend"
version = "0.1.0"
description = "Flask backend for HR SASS project"
authors = ["Your Name <you@example.com>"]
packages = [{ include = "hr_backend" }]

[tool.poetry.dependencies]
python = "^3.11"
flask = "^2.3.0"
flask-cors = "^4.0.0"
sqlalchemy = "^2.0"
psycopg = { version = "^3.2", extras = ["binary"] }
psycopg2-binary = "^2.9"
PyJWT = "^2.7"
bcrypt = "^4.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
alembic = "^1.11"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
